move reporting-issue related classes to package `feedback`
