+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/feedback/NewCustomerIssueFeedbackAction.java
@@ -3,13 +3,13 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-package com.microsoft.azure.toolkit.intellij.common.action;
+package com.microsoft.azure.toolkit.intellij.feedback;
 
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.microsoft.azuretools.telemetry.TelemetryConstants;
-import com.microsoft.intellij.feedback.GithubIssue;
-import com.microsoft.intellij.feedback.NewGithubIssueAction;
-import com.microsoft.intellij.feedback.ReportableIssue;
+import com.microsfot.azure.toolkit.intellij.feedback.GithubIssue;
+import com.microsfot.azure.toolkit.intellij.feedback.NewGithubIssueAction;
+import com.microsfot.azure.toolkit.intellij.feedback.ReportableIssue;
 
 public class NewCustomerIssueFeedbackAction extends NewGithubIssueAction {
     public NewCustomerIssueFeedbackAction() {
