+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/feedback/NewFeatureRequestFeedbackAction.java
@@ -3,18 +3,18 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-package com.microsoft.azure.toolkit.intellij.common.action;
+package com.microsoft.azure.toolkit.intellij.feedback;
 
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.microsoft.azuretools.telemetry.TelemetryConstants;
-import com.microsoft.intellij.feedback.GithubIssue;
-import com.microsoft.intellij.feedback.NewGithubIssueAction;
-import com.microsoft.intellij.feedback.ReportableFeatureRequest;
+import com.microsfot.azure.toolkit.intellij.feedback.GithubIssue;
+import com.microsfot.azure.toolkit.intellij.feedback.NewGithubIssueAction;
+import com.microsfot.azure.toolkit.intellij.feedback.ReportableFeatureRequest;
 
 public class NewFeatureRequestFeedbackAction extends NewGithubIssueAction {
     public NewFeatureRequestFeedbackAction() {
         super(new GithubIssue<>(new ReportableFeatureRequest("Feature Request"))
-                .withLabel("feature-request"), "Request Features");
+            .withLabel("feature-request"), "Request Features");
     }
 
     @Override
