+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/kotlin/com/microsfot/azure/toolkit/intellij/feedback/NewGithubIssueAction.kt
@@ -20,12 +20,15 @@
  * SOFTWARE.
  */
 
-package com.microsoft.intellij.feedback
+package com.microsfot.azure.toolkit.intellij.feedback
 
-class ReportableError(shortMessage: String, private val detailMessage: String)
-            : Reportable(shortMessage){
-    override val detailInfo
-        get() = mapOf(
-                "Error message" to GithubMarkdownFormat.toCode(detailMessage)
-        )
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.microsoft.azure.toolkit.intellij.common.action.AzureAnAction
+import com.microsoft.azuretools.telemetrywrapper.Operation
+
+open class NewGithubIssueAction(private val issue: GithubIssue<out Reportable>, actionTitle: String) : AzureAnAction(actionTitle) {
+    override fun onActionPerformed(anActionEvent: AnActionEvent, operation: Operation?): Boolean {
+        issue.report()
+        return true
+    }
 }
