+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/kotlin/com/microsfot/azure/toolkit/intellij/feedback/ReportableSurvey.kt
@@ -20,14 +20,16 @@
  * SOFTWARE.
  */
 
-package com.microsoft.intellij.feedback
+package com.microsfot.azure.toolkit.intellij.feedback
 
-object GithubMarkdownFormat {
-    fun toCode(value: String): String {
-        return "\n```\n$value\n```"
-    }
+class ReportableSurvey(shortMessage: String) : Reportable(shortMessage) {
+    private val questions = listOf(
+        "#### 1. Are you satisfied with the Apache Spark development functionalities & experiences offered in Azure Toolkit for IntelliJ? Please rate us from 1 (poor) to 10 (excellent).",
+        "#### 2. What did you like or dislike about Azure Toolkit for IntelliJ?",
+        "#### 3. What enhancements would you like Azure Toolkit for IntelliJ to implement?"
+    )
 
-    fun toInlineCode(value: String): String {
-        return "`$value`"
+    override fun getBody(): String {
+        return super.getBody() + "\n\n" + questions.joinToString("\n\n") + "\n\n"
     }
 }
